+++ apps/desktop/src/app/session/hooks/use-prompt-actions/utils.ts
@@ -4,6 +4,7 @@ import { translateNow, type Translations } from '@/i18n'
 import type { ChatMessage } from '@/lib/chat-messages'
 import { type CommandsCatalogLike, filterDesktopCommandsCatalog } from '@/lib/desktop-slash-commands'
 import { isProviderSetupErrorMessage } from '@/lib/provider-setup-errors'
+import type { ComposerAttachment } from '@/store/composer'
 
 export type GatewayRequest = <T>(method: string, params?: Record<string, unknown>) => Promise<T>
 
@@ -208,4 +209,9 @@ export function visibleUserIndexAtOrdinal(messages: readonly ChatMessage[], targ
   }
 
   return -1
+}
+
+export interface SubmitTextOptions {
+  attachments?: ComposerAttachment[]
+  fromQueue?: boolean
 }
